Restructure the slate-hyperscript; I cannot understand and develop on a large all-in-one file
