+++ packages/slate-hyperscript/src/constants/selection.js
@@ -1,0 +1,11 @@
+/**
+ * Create selection point constants, for comparison by reference.
+ *
+ * @type {Object}
+ */
+
+const ANCHOR = {}
+const CURSOR = {}
+const FOCUS = {}
+
+export { ANCHOR, CURSOR, FOCUS }
